@@ -6,21 +6,30 @@
     "keywords": ["cakephp", "jwt", "cakephp jwt", "cakephp jwt auth", "cakephp jwt authentication"],
     "require": {
         "ext-openssl": "*",
-        "php": "^8.0",
-        "cakephp/authentication": "2.9.*",
-        "cakephp/cakephp": "^4.2",
-        "firebase/php-jwt": "^5.5"
+        "php": "^8.1",
+        "cakephp/authentication": "^3.0",
+        "cakephp/cakephp": "^5.0",
+        "firebase/php-jwt": "^6.2"
     },
     "require-dev": {
-        "phpunit/phpunit": "^8.5 || ^9.3"
+        "josegonzalez/dotenv": "^3.2",
+        "phpunit/phpunit": "^10.0"
     },
     "autoload": {
         "psr-4": {
             "MixerApi\\JwtAuth\\": "src/"
         }
     },
+    "autoload-dev": {
+        "psr-4": {
+            "MixerApi\\JwtAuth\\": "plugins/jwt-auth/src/",
+            "MixerApi\\JwtAuth\\Test\\": "plugins/jwt-auth/tests/",
+            "MixerApi\\JwtAuth\\Test\\App\\": "plugins/jwt-auth/tests/test_app/src/"
+        }
+    },
     "support": {
-        "source": "https://github.com/mixerapi/jwt-auth"
+        "issues": "https://github.com/mixerapi/mixerapi-dev/issues",
+        "source": "https://github.com/mixerapi/mixerapi-dev/"
     },
     "authors": [
         {
@@ -29,5 +38,11 @@
         }
     ],
     "prefer-stable": true,
-    "minimum-stability": "dev"
+    "minimum-stability": "dev",
+    "config": {
+        "allow-plugins": {
+            "dealerdirect/phpcodesniffer-composer-installer": true,
+            "cakephp/plugin-installer": true
+        }
+    }
 }
